requirements.txt
fastapi
uvicorn
gunicorn
jinja2
pytest
requests
httpx
pre-commit
python-dotenv
pydantic
pydantic-settings
python-multipart
Pillow
pytesseract
tesseract --list-langs

